# pied_piper_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pied Piper",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reaction/toggle/{type}/{id}": {
            "post": {
                "tags": [
                    "reactions"
                ],
                "summary": "create/delete/updeate comment",
                "description": "User create/delete/updeate reactions here",
                "operationId": "create/delete/updeate_reaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "the id of post/comment you want create reaction on",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "type": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "reaction created successfully!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
        "/api/users": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "get users by search",
                "operationId": "search for users",
                "parameters": [
                    {
                        "name": "searrch",
                        "in": "query",
                        "description": "enter name of user you want to search for",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all found users"
                    },
                    "401": {
                        "description": "unauthorized."
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/search/add/{id}": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "add to history",
                "operationId": "addToHistory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "enter the user searched id to add",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "added to search history",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/search/delete/{id}": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "delete from search history",
                "operationId": "deleteFromHistory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "enter the user searched id to remove from search history",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "removed from search history",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/search/all": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "get users by search",
                "operationId": "getSearchHistory",
                "responses": {
                    "200": {
                        "description": "all search history"
                    },
                    "401": {
                        "description": "unauthorized."
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}